Fix using dynamic boost libraries and streamline project-config generation
Fix the building and linking of dynamic boost libraries that depend on
setting a BOOST_<LIB>_DYN_LINK macro. Ensure that the macro is correctly
added both when building the library but also when building dependent
application code that will be linked to the library.

Also make libraries only Require() bjam and project-config files so that
bjam is not re-executed every time for libraries.