Update boost test runner to work with vanilla boost (against all odds)
This is a first pass at this and works for most cases but there may be
aperformance hit if used with many assertions since this relies on
logging every assertion in order to figure out what being called. This
is needed because of boost.tests completely broken approach to logging.
The original boost test runner is changed to "patched_boost" for those
lucky enough to have the time and effort to get that set up.