GCC 5 support, add cxx_version to toolchain, add suppress -g for auto in clang
cxx_version allows the binary version for calling to be different from
the reported version. For example on Debian to execute GCC 5 you need to
call g++-5 but the actual version (at time of writing) is 5.1. This
makes it possible to write a correct project config jam for boost and so
on. It also make it possible to specify gcc51 on the command line as a
toolchain but have cuppa execute g++-5 knowing that in fact that will be
the correct binary for g++ 5.1.