Second take at BuildSharedLib fix.
The BuildLib method creates a static or shared library depending on
a 'shared' argument which defaults to False.
The BuildStaticLib method calls this with shared=False.
The BuildSharedLib method calls this with shared=True.

Similarly, the Compile method creates a static or shared object
depending on a 'shared' argument which defaults to False.
The CompileStatic method calls this with shared=False.
The CompileShared method calls this with shared=True.